refactor: replace eval with repl where possible

+++ crates/nixlike/src/se_impl.rs
@@ -212,7 +212,7 @@ impl Serializer for MySerialize {
 	}
 
 	fn serialize_i64(self, v: i64) -> Result<Self::Ok, Self::Error> {
-		Ok(Value::Number(v as i64))
+		Ok(Value::Number(v))
 	}
 
 	fn serialize_u8(self, v: u8) -> Result<Self::Ok, Self::Error> {
